move extractor to utils and add tests
