+++ dspy/adapters/json_adapter.py
@@ -10,6 +10,7 @@ from pydantic.fields import FieldInfo
 from dspy.adapters.chat_adapter import ChatAdapter, FieldInfoWithName
 from dspy.adapters.types.tool import ToolCalls
 from dspy.adapters.utils import (
+    _extract_first_json_object,
     format_field_value,
     get_annotation_name,
     parse_value,
@@ -295,42 +296,3 @@ def _get_structured_outputs_response_format(
     pydantic_model.model_json_schema = lambda *args, **kwargs: schema
 
     return pydantic_model
-
-def _extract_first_json_object(text: str) -> str | None:
-    """Return the first balanced JSON object found in text or None if absent."""
-
-    in_string = False
-    escape = False
-    depth = 0
-    start_idx: int | None = None
-    seen_lbrace = False
-
-    for idx, char in enumerate(text):
-        if seen_lbrace and in_string:
-            if escape:
-                escape = False
-            elif char == "\\":
-                escape = True
-            elif char == '"':
-                in_string = False
-            continue
-
-        if seen_lbrace and char == '"':
-            in_string = True
-            continue
-
-        if char == "{":
-            if depth == 0:
-                start_idx = idx
-                seen_lbrace = True
-            depth += 1
-            continue
-
-        if char == "}":
-            if depth == 0 or start_idx is None:
-                continue
-            depth -= 1
-            if depth == 0:
-                return text[start_idx : idx + 1]
-
-    return None
